Update type-inference of annotated expressions
… as standardized in https://github.com/dhall-lang/dhall-lang/pull/745.

This also adds proper error handling for the free variable case
discussed in https://github.com/dhall-lang/dhall-lang/issues/749:

Sample output for the MergeHandlerFreeVar test case:

    Error: Disallowed handler type

    Explanation: You can ❰merge❱ the alternatives of a union using a record with one
    handler per alternative, like this:

        ┌─────────────────────────────────────────────────────────────────┐
        │ let union    = < Left : Natural | Right : Bool >.Left 2         │
        │ let handlers = { Left = Natural/even, Right = λ(x : Bool) → x } │
        │ in  merge handlers union : Bool                                 │
        └─────────────────────────────────────────────────────────────────┘

    ... but the output type of a handler may not depend on the input value.

    For example, the following expression is not valid:

          Invalid: The output type is ❰Optional A❱, which references the input
          value ❰A❱.
                      ⇩
        ┌──────────────────────────────────────────┐
        │ merge { x = None } (< x : Type >.x Bool) │
        └──────────────────────────────────────────┘

    Your handler for the following alternative:

    ↳ x

    ... has type:

    ↳ ∀(A : Type) → Optional A

    ... where the output type:

    ↳ Optional A

    ... references the handler's input value:

    ↳ A

    ────────────────────────────────────────────────────────────────────────────────

    1│ merge { x = None } (<x: Type>.x Bool)

    dhall/dhall-lang/tests/type-inference/failure/unit/MergeHandlerFreeVar.dhall:1:1

+++ dhall/src/Dhall/Eval.hs
@@ -975,7 +975,7 @@ quote !env !t0 =
         VConst k ->
             Const k
         VVar !x !i ->
-            Var (V x (fromIntegral (countNames x env - i - 1)))
+            Var (V x (countNames x env - i - 1))
         VApp t u ->
             quote env t `qApp` u
         VLam a (freshClosure -> (x, v, t)) ->
